requirements.txt
asgiref==3.2.10
Django==3.0.3
django-appconf==1.0.4
django-crispy-forms==1.8.1
django-imagefit==0.6.2
django-mathfilters==1.0.0
djangorestframework==3.11.0
html5lib==1.1
mod-wsgi==4.7.1
module-name==0.5.1
mysqlclient==1.4.2.post1
Pillow==7.0.0
PyPDF2==1.26.0
python-dateutil==2.8.1
pytz==2020.1
reportlab==3.5.52
RouterOS-api==0.17.0
six==1.15.0
sqlparse==0.3.1
webencodings==0.5.1
xhtml2pdf==0.2.4
